+++ src/router/index.js
@@ -2,7 +2,10 @@ import { createRouter, createWebHistory } from "vue-router";
 import Main from '../components/Home'
 import Vaccine from '../components/vaccine/Main'
 import Crane from '../components/crane/CraneMain'
-import Elecar from '../components/elecar/ElecarMain'
+import Elecar from '../components/elecar/Elecar'
+import ElecarMain from '../components/elecar/ElecarMain'
+import ElecarList from '../components/elecar/ElecarList'
+
 const routes = [
     {
         path : '/',
@@ -22,7 +25,19 @@ const routes = [
     {
         path : '/elecar',
         name : 'elecar',
-        component : Elecar
+        component : Elecar,
+        children: [
+            {
+                path : '',
+                name : 'main',
+                component : ElecarMain
+            },
+            {
+                path : 'list',
+                name : 'elecarList',
+                component : ElecarList
+            }
+        ]
     },
 ]
